Adjust Component File Structure
